feat: 체크인 인증 엔드포인트 및 관련 DTO 추가 - 방문 인증을 생성하고 보상(EXP, 레벨, 스탬프, 배지)을 반환하는 /certifications/checkin 엔드포인트를 구현함 - 요청/응답 구조를 위해 dto.CreateCheckInRequest, dto.CertificationResponse 추가 - 보상 및 배지 처리를 위해 dto.NewBadgeResponse, dto.RewardResponse, dto.StampResponse 도입 - 경험치 계산, 레벨 갱신, 배지 지급을 포함하도록 GamificationService 로직 확장 - 매장별 사용자 활동을 추적하기 위한 UserStoreStat 엔티티 생성 - 신규 엔드포인트에 대한 Swagger 문서 보강 및 기존 DTO 예제 업데이트

+++ certifications/entities/certification.go
@@ -8,9 +8,9 @@ import (
 )
 
 type Certification struct {
-	ID         uint      `gorm:"primaryKey;autoIncrement" json:"id"`
-	UserID     uint      `gorm:"column:userId;not null;index:idx_cert_user_time" json:"userId"`
-	StoreID    uint      `gorm:"column:storeId;not null;index:idx_cert_store_time" json:"storeId"`
+	ID         int64     `gorm:"primaryKey;autoIncrement" json:"id"`
+	UserID     int64     `gorm:"column:userId;not null;index:idx_cert_user_time" json:"userId"`
+	StoreID    int64     `gorm:"column:storeId;not null;index:idx_cert_store_time" json:"storeId"`
 	Type       string    `gorm:"type:enum('loot','checkin');not null" json:"type"`              // loot:득템, checkin:빈손방문
 	OccurredAt time.Time `gorm:"column:occurredAt;type:datetime(6);not null" json:"occurredAt"` // 실제 방문 시간
 	Latitude   *float64  `gorm:"type:decimal(10,6)" json:"latitude"`                            // 인증 당시 GPS
@@ -21,12 +21,12 @@ type Certification struct {
 	CreatedAt  time.Time `gorm:"column:createdAt;type:datetime(6);autoCreateTime" json:"createdAt"`
 
 	// Associations
-	User    *users.User              `gorm:"foreignKey:UserID;constraint:OnDelete:CASCADE" json:"user,omitempty"`
-	Store   *stores.Store            `gorm:"foreignKey:StoreID;constraint:OnDelete:CASCADE" json:"store,omitempty"`
-	Photos  []CertificationPhoto     `gorm:"foreignKey:CertificationID" json:"photos,omitempty"`
-	Tags    []users.LootTag          `gorm:"many2many:certification_tags" json:"tags,omitempty"`
-	Reasons []CheckinReasonPreset    `gorm:"many2many:certification_reasons" json:"reasons,omitempty"`
-	Likes   []LootLike               `gorm:"foreignKey:CertificationID" json:"likes,omitempty"`
+	User    *users.User           `gorm:"foreignKey:UserID;constraint:OnDelete:CASCADE" json:"user,omitempty"`
+	Store   *stores.Store         `gorm:"foreignKey:StoreID;constraint:OnDelete:CASCADE" json:"store,omitempty"`
+	Photos  []CertificationPhoto  `gorm:"foreignKey:CertificationID" json:"photos,omitempty"`
+	Tags    []users.LootTag       `gorm:"many2many:certification_tags" json:"tags,omitempty"`
+	Reasons []CheckinReasonPreset `gorm:"many2many:certification_reasons" json:"reasons,omitempty"`
+	Likes   []LootLike            `gorm:"foreignKey:CertificationID" json:"likes,omitempty"`
 }
 
 func (Certification) TableName() string {
